Remove definitions/interface for corruptedFrames.
Media WG resolved to remove this field at TPAC.
https://www.w3.org/2019/09/19-mediawg-minutes.html#resolution01

Most UAs have not implemented corrupted frames and have no means to do
so.

Serious corruption will generally produce a decode error. This may
may not be logically associated with particular frames. Less serious
corruption may produce artifacts in the output, but in most cases it is
not feasible for a UA to know that this has occurred.